pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "carlos47fnaf", email = "kaduardomello@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "djangorestframework>=3.14.0",
    "setuptools>=68.1.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
